refactor: customcard widget
