+++ lib/views/widgets/custom_card.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+
+class CustomCard extends StatelessWidget {
+  final String title;
+  final IconData icon;
+  final VoidCallback onTap;
+  CustomCard({required this.title, required this.icon, required this.onTap});
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      margin: const EdgeInsets.symmetric(vertical: 10, horizontal: 25),
+      child: InkWell(
+        onTap: onTap,
+        child: ListTile(
+          leading: FaIcon(icon),
+          title: Text(title),
+        ),
+      ),
+    );
+  }
+}
